requirements.txt
anyio==3.6.2
bcrypt==4.0.1
click==8.1.3
dnspython==2.2.1
email-validator==1.3.0
fastapi==0.86.0
greenlet==2.0.1
h11==0.14.0
idna==3.4
psycopg2-binary==2.9.5
pydantic==1.10.2
python-dotenv==0.21.0
sniffio==1.3.0
SQLAlchemy==1.4.43
starlette==0.20.4
typing_extensions==4.4.0
uvicorn==0.19.0
